Title: 체질량지수, Time: 128ms, Memory: 10MB, Status: Wrong Answer - CodeTree

+++ 240617/체질량지수/bmi.java
@@ -6,7 +6,7 @@ public class Main {
 
         Scanner sc = new Scanner(System.in);
 
-        int h, w;
+        long h, w;
 
         h = sc.nextInt();
         w = sc.nextInt();
